requirements.txt
annotated-types==0.7.0
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.4
django-appconf==1.0.6
django-cors-headers==4.3.1
django-ninja==1.1.0
django-vite==3.0.4
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
pydantic==2.8.2
pydantic-core==2.20.1
PyJWT==2.6.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2022.7
rcssmin==1.1.1
requests==2.28.1
requests-oauthlib==1.3.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.13
whitenoise==6.6.0
celery==5.4.0
